AFL_PROJECT/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
